Ignore dotfiles.
If the file or directory name starts with a dot, skip it.

Perhaps not what we eventually want for a scrubber, but
makes testing simpler at this stage.

The `if let` clause is a bit clunky, but have to deal with
multiple possible conversion errors. Maybe .to_string_lossy()
would be more appropriate for this sort of task?