Recurse into subdirectories.
Just call ourselves recursively when we find a directory, hoping
there are no loops.

Shift the argument type from &str to &Path so we can recurse
directly on what we get from the direntry. Converting paths
to strings can fail unicode validation, so while it's convenient
to pass in as the initial directory it's not appropriate for
handling data which came from the filesystem.